dev-requirements.txt
opencv-python
numpy==1.23.5
matplotlib==3.7
roboticstoolbox-python
pandas
